Solution for day 7
This one was highly delayed by an attempt to use Box. Namely, using Box<str> instead of String. Eventually I gave in to an extra usize on the stack for non-changing strings.

I considered a hash map for Part 1, but figured I didn't need it
performance-wise. So Part 2 reconstructs into a hashmap.